requirements.txt
hydra-core
numpy
torch
pandas
SoccerNet
rich
ultralytics
SportsLabKit